Various message passing improvements
Messages are now more simple, versatile, and efficient. Each message
represents
a length-prefixed single shm variable, which makes packetizing messages
simpler. Different types of var values are now a oneof instead of
optionals which saves memory, and a lack of a value represents a variable read request.
After receiving a read request, the copter will send back a variable
value message of the specified variable. In the future, this technique
could be used to easily communicate between multiple microcontrollers
with shms of their own.

The client now recognizes the Bluetooth LE 20-byte packet limit and will
attempt to pack as many shm messages as possible into a single packet,