Exclusively use serial for all communication
Now that wireless communication is in the hands of the handheld
controller (pun half intended), the client only needs to read and write
raw serial data and doesn't need to use Bluetooth nor encode control
desire inputs. This provides an elegant abstraction which allows the
client to either communicate with the drone through the handheld (and
thus through radio), or directly with the drone over serial if it is
connected this way.